// src/app/model/viewtask.ts
export class ViewTask{
    taskId: number;
    taskDesc: string;
    startDate: string;
    endDate: string;
    priority: number;
    taskStatus: boolean;
    parentId: number;
    parentTask: string;
    projectId: number;
    projectName: string;
    projectStartDate: string;
    projectEndDate: string;
    projectPriority: number;
    projectStatus: boolean;
    userId: number;
    firstName: string;
    lastName: string;
    employeeId: number;
}