// src/app/model/project.ts
export class Project {
    projectId: number;
    projectDesc: string;
    projectStartDate: string;
    projectEndDate: string;
    projectPriority: number;
    numberOfTasks: number;
    completedTasks: number;
    userId: number;
    firstName: string;
    lastName: string;
    employeeId: number;
    projectStatus: boolean;
}